Put CRD openapi controller in aggregator apiserver
to speed publishing up
